Update PebbleKit to SDK 2.0
It should work with both firmware 1.x and 2.x (tested with fw 1.13.0 and 2.0-beta2)
It allows to use 2.x firmware functions (such as getWatchFWVersion), if Exception are correctly catched on firmware 1.x

https://developer.getpebble.com/2/mobile-app-guide/android-guide.html/

PebbleKit Android is distributed in source form as an Android project. Before you can use it in a project, you’ll need to load it into your IDE.

To load it into Eclipse, follow these simple steps:

    Click File –> New Project…
    Select Android Project from Existing Code – Press Next
    Select ~/pebble-dev/PebbleSDK-2.0-BETA2/PebbleKit-Android/PebbleKit as the root directory
    Click Finish

This loadd a new project called PEBBLE_KIT. Ensure there are no errors on this project.

To use PebbleKit Android in one of your Android project:

    Right-click your project and select Properties
    Select the Android pane
    Click Add…
    Select PEBBLE_KIT

That’s it! You can now use PebbleKit in your project.