Send last saved value when the gps is started + fix tests
ArgumentCaptor.forClass(GPSStatus.class) DOEST NOT check for GPSStatus type... It is just a cast and you can get any class in the capture...

http://docs.mockito.googlecode.com/hg/1.9.5/org/mockito/ArgumentCaptor.html#forClass(java.lang.Class)
Note that an ArgumentCaptor *don't do any type checks*, it is only there to avoid casting in your code. This might however change (type checks could be added) in a future major release.